// src/app/models/user.ts
'use strict';

export class User {
    userId: string;
    userName: string;
    fullName: string;
    dob: Date;
    email: string;
    userRole: number = 2;
    userImage: string;
    teamName: string;
    teamImage: string;
    teamMotto: string;
    supportingCountry: number;
    password: string;
}